passed make compilation. Next need to debug the code

+++ packet_handler.c
@@ -33,9 +33,20 @@ char *build_whohas_query(vector *chunks_to_download) {
     return query;
 }
 
-handler_input *build_handler_input(uint16_t incoming_socket, char *body_buf,
+/**
+ * build the handler input struct
+ * @param incoming_socket
+ * @param body_buf
+ * @param from_ip
+ * @param from_len
+ * @param buf_len
+ * @param recv_size
+ * @param header
+ * @return
+ */
+handler_input *build_handler_input(int incoming_socket, char *body_buf,
                                    struct socket_in *from_ip, socklen_t
-                                   from_len, uint16_t buf_len, uint32_t
+                                   from_len, int buf_len, int
                                    recv_size, packet_h *header) {
     handler_input *res = (handler_input *) malloc(sizeof(handler_input));
     res->incoming_socket = incoming_socket;
